move the delete all mandatory training files into a new folder
