+++ src/app/features/add-mandatory-training/add-mandatory-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { AddAndManageMandatoryTrainingComponent } from './add-and-manage-mandatory-training/add-and-manage-mandatory-training.component';
 import { AddMandatoryTrainingComponent } from './add-mandatory-training.component';
-import { RemoveAllMandatoryTrainingComponent } from './delete-all-mandatory-training.component';
+import { RemoveAllMandatoryTrainingComponent } from './delete-mandatory-training/delete-all-mandatory-training.component';
 
 const routes: Routes = [
   {
