+++ src/app/features/add-mandatory-training/add-mandatory-training.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { AddMandatoryTrainingRoutingModule } from '@features/add-mandatory-training/add-mandatory-routing.module';
-import { RemoveAllMandatoryTrainingComponent } from '@features/add-mandatory-training/delete-all-mandatory-training.component';
+import { RemoveAllMandatoryTrainingComponent } from '@features/add-mandatory-training/delete-mandatory-training/delete-all-mandatory-training.component';
 import { SharedModule } from '@shared/shared.module';
 
 import { AddAndManageMandatoryTrainingComponent } from './add-and-manage-mandatory-training/add-and-manage-mandatory-training.component';
