+++ src/app/features/add-mandatory-training/delete-mandatory-training/delete-all-mandatory-training.component.ts
